Move code into components/device-sync.js
